+++ enjoy/src/main/db/handlers/index.ts
@@ -3,7 +3,6 @@ export * from "./cache-objects-handler";
 export * from "./chat-agents-handler";
 export * from "./chat-members-handler";
 export * from "./chat-messages-handler";
-export * from "./chat-sessions-handler";
 export * from "./chats-handler";
 export * from "./conversations-handler";
 export * from "./messages-handler";
