+++ enjoy/src/renderer/components/chats/index.ts
@@ -5,5 +5,4 @@ export * from "./chat-card";
 export * from "./chat-form";
 export * from "./chat-input";
 export * from "./chat-message";
-export * from "./chat-sessions";
 export * from "./chat-sidebar";
