+++ enjoy/src/renderer/hooks/index.ts
@@ -3,7 +3,7 @@ export * from "./use-audio";
 export * from "./use-camdict";
 export * from "./use-chat";
 export * from "./use-chat-agent";
-export * from "./use-chat-session";
+export * from "./use-chat-message";
 export * from "./use-conversation";
 export * from "./use-notes";
 export * from "./use-recordings";
